+++ bpdm-pool-api/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/api/client/PoolApiClient.kt
@@ -35,5 +35,7 @@ interface PoolApiClient {
 
     val members: MembersApiClient
 
+    val memberOwned: MemberOwnedApiClient
+
     val memberships: CxMembershipApiClient
 }