+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/AddressService.kt
@@ -28,7 +28,6 @@ import org.eclipse.tractusx.bpdm.pool.api.model.LogisticAddressVerboseDto
 import org.eclipse.tractusx.bpdm.pool.entity.LogisticAddressDb
 import org.eclipse.tractusx.bpdm.pool.repository.LegalEntityRepository
 import org.eclipse.tractusx.bpdm.pool.repository.LogisticAddressRepository
-import org.eclipse.tractusx.bpdm.pool.repository.SiteRepository
 import org.springframework.data.domain.PageRequest
 import org.springframework.data.jpa.domain.Specification
 import org.springframework.stereotype.Service
@@ -36,8 +35,7 @@ import org.springframework.stereotype.Service
 @Service
 class AddressService(
     private val logisticAddressRepository: LogisticAddressRepository,
-    private val legalEntityRepository: LegalEntityRepository,
-    private val siteRepository: SiteRepository,
+    private val legalEntityRepository: LegalEntityRepository
 ) {
     private val logger = KotlinLogging.logger { }
 
@@ -52,7 +50,8 @@ class AddressService(
             LogisticAddressRepository.bySiteBpns(searchRequest.siteBpns),
             LogisticAddressRepository.byLegalEntityBpns(searchRequest.legalEntityBpns),
             LogisticAddressRepository.byName(searchRequest.name),
-            LogisticAddressRepository.byIsMember(searchRequest.isCatenaXMemberData)
+            LogisticAddressRepository.byIsMember(searchRequest.isCatenaXMemberData),
+            LogisticAddressRepository.byIsMemberOwned(searchRequest.isMemberOwnedData),
         )
         val addressPage = logisticAddressRepository.findAll(spec, paginationRequest.toPageRequest())
 
@@ -92,6 +91,7 @@ class AddressService(
         val siteBpns: List<String>?,
         val legalEntityBpns: List<String>?,
         val name: String?,
-        val isCatenaXMemberData: Boolean?
+        val isCatenaXMemberData: Boolean?,
+        val isMemberOwnedData: Boolean?
     )
 }